BookStore/requirements.txt
asgiref==3.7.2
distlib==0.3.7
Django==4.2.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.13.1
platformdirs==3.11.0
psycopg2==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
sqlparse==0.4.4
tzdata==2023.3
virtualenv==20.24.6
